requirements.txt
psycopg2-binary
django-import-export
djangorestframework
django-cors-headers
Django==6.0.2
python-dotenv
whitenoise
waitress
Pillow
